# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HistoryTalk API",
    "version": "1.0.0",
    "description": "Vietnamese history educational AI platform — backend REST API"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Local development"
    },
    {
      "url": "https://api.historytalk.dev/api/v1",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/characters": {
      "get": {
        "tags": [
          "Characters"
        ],
        "summary": "List characters"
      },
      "post": {
        "tags": [
          "Characters"
        ],
        "summary": "Create a character",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/characters/context/{contextId}": {
      "get": {
        "tags": [
          "Characters"
        ],
        "summary": "List characters by context ID",
        "parameters": [
          {
            "in": "path",
            "name": "contextId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/characters/{id}": {
      "get": {
        "tags": [
          "Characters"
        ],
        "summary": "Get character by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "put": {
        "tags": [
          "Characters"
        ],
        "summary": "Update a character",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "Characters"
        ],
        "summary": "Delete a character",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/characters/{id}/soft-delete": {
      "patch": {
        "tags": [
          "Characters"
        ],
        "summary": "Soft-delete a character",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/characters/{id}/toggle-active": {
      "patch": {
        "tags": [
          "Characters"
        ],
        "summary": "Toggle character active status",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/characters/{characterId}/contexts/{contextId}": {
      "post": {
        "tags": [
          "Characters"
        ],
        "summary": "Attach a character to a historical context",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "characterId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "contextId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "Characters"
        ],
        "summary": "Remove a character from a historical context",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "characterId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "contextId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/characters/{characterId}/contexts": {
      "get": {
        "tags": [
          "Characters"
        ],
        "summary": "Get all contexts of a character",
        "parameters": [
          {
            "in": "path",
            "name": "characterId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/system-admin/dashboard/quiz": {
      "get": {
        "tags": [
          "System Admin Dashboard"
        ],
        "summary": "Get quiz analytics",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/historical-contexts": {
      "get": {
        "tags": [
          "Historical Contexts"
        ],
        "summary": "List historical contexts"
      },
      "post": {
        "tags": [
          "Historical Contexts"
        ],
        "summary": "Create a historical context",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/historical-contexts/{id}": {
      "get": {
        "tags": [
          "Historical Contexts"
        ],
        "summary": "Get historical context by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "put": {
        "tags": [
          "Historical Contexts"
        ],
        "summary": "Update a historical context",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "Historical Contexts"
        ],
        "summary": "Delete a historical context",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/historical-contexts/{id}/soft-delete": {
      "patch": {
        "tags": [
          "Historical Contexts"
        ],
        "summary": "Soft-delete a historical context",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/historical-contexts/{id}/toggle-active": {
      "patch": {
        "tags": [
          "Historical Contexts"
        ],
        "summary": "Toggle historical context active status",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/quizzes/results/me": {
      "get": {
        "tags": [
          "Quizzes"
        ],
        "summary": "Get my quiz results",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/quizzes/submit": {
      "post": {
        "tags": [
          "Quizzes"
        ],
        "summary": "Submit quiz answers",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/quizzes/{quizId}": {
      "get": {
        "tags": [
          "Quizzes"
        ],
        "summary": "Get quiz by ID",
        "parameters": [
          {
            "in": "path",
            "name": "quizId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/quizzes/{quizId}/start": {
      "post": {
        "tags": [
          "Quizzes"
        ],
        "summary": "Start a quiz session",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "quizId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
...
}